cli implementation
