+++ Cargo.toml
@@ -31,6 +31,7 @@ async-stream = "0.3.5"
 bincode = "1.3"
 cairo-proof-parser = { git = "https://github.com/Okm165/cairo-proof-parser", rev = "97a04bbee07330311b38d6f4cecfed3acb237626" }
 cairo-vm = { git = "https://github.com/lambdaclass/cairo-vm.git", tag = "v1.0.0-rc3" }
+clap = { version = "4.4.4", features = ["derive"] }
 futures = "0.3.30"
 futures-core = "0.3.30"
 futures-util = "0.3.30"
