+++ crates/common/Cargo.toml
@@ -12,6 +12,7 @@ publish = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+clap.workspace = true
 cairo-vm.workspace = true
 rand.workspace = true
 futures.workspace = true
