+++ crates/common/src/network.rs
@@ -3,10 +3,11 @@
     This defines the network that the node is connected to.
 */
 
+use clap::ValueEnum;
 use starknet::core::chain_id;
 use starknet_crypto::FieldElement;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, ValueEnum, Clone, Copy)]
 pub enum Network {
     Mainnet,
     Sepolia,
