+++ crates/delegator/Cargo.toml
@@ -8,9 +8,14 @@ readme = "README.md"
 repository.workspace = true
 version.workspace = true
 
+[[bin]]
+name = "zetina-submit"
+path = "src/main.rs"
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+clap.workspace = true
 futures.workspace = true
 hex.workspace = true
 libp2p.workspace = true
